[DoctrineBridge] Changed UniqueEntityValidator to use the 2.5 Validation API
