+++ src/Symfony/Bridge/Doctrine/Validator/Constraints/UniqueEntityValidator.php
@@ -13,6 +13,7 @@ namespace Symfony\Bridge\Doctrine\Validator\Constraints;
 
 use Doctrine\Common\Persistence\ManagerRegistry;
 use Symfony\Component\Validator\Constraint;
+use Symfony\Component\Validator\Context\ExecutionContextInterface;
 use Symfony\Component\Validator\Exception\UnexpectedTypeException;
 use Symfony\Component\Validator\Exception\ConstraintDefinitionException;
 use Symfony\Component\Validator\ConstraintValidator;
@@ -136,6 +137,14 @@ class UniqueEntityValidator extends ConstraintValidator
 
         $errorPath = null !== $constraint->errorPath ? $constraint->errorPath : $fields[0];
 
-        $this->context->addViolationAt($errorPath, $constraint->message, array(), $criteria[$fields[0]]);
+        if ($this->context instanceof ExecutionContextInterface) {
+            $this->context->buildViolation($constraint->message)
+                          ->atPath($errorPath)
+                          ->setInvalidValue($criteria[$fields[0]])
+                          ->addViolation();
+        } else {
+            // 2.4 API
+            $this->context->addViolationAt($errorPath, $constraint->message, array(), $criteria[$fields[0]]);
+        }
     }
 }
